+++ .gitignore
@@ -27,3 +27,5 @@ scr/functions/__pycache__/classes.cpython-310.pyc
 scr/functions/__pycache__/test_func.cpython-310-pytest-7.1.2.pyc
 scr/functions/__pycache__/tools.cpython-310.pyc
 scr/tests/__pycache__/test_func.cpython-310-pytest-7.1.2.pyc
+scr/arguments/__pycache__/arguments.cpython-310.pyc
+scr/arguments/__pycache__/__init__.cpython-310.pyc
