+++ scr/functions/__init__.py
@@ -1,6 +1,5 @@
 # trunk-ignore(flake8/F401)
 from functions.tools import (
-    argument,
     blue,
     create_config,
     getInput,
